Fix bugs and add kind of cache ...
Fix editable information (when we are in automatic mode).
Fix missing display name when creating a document for url.
Use synchronization mechanism for contacts retrieval: list is updated.
It add correctly new contacts,
but there is a strange behavior when merging contacts.
Or deleting one.
I guess that every contact which index has been changed are sent.